Stop replies to commands from appearing in the GUI before the commands themselves
The reason this was happening is because the replies were being sent during
the handler for the client's message, and the acknowledgement of the client's
message was only sent *after* the handler finished. Sending the replies from
a separate goroutine fixes this in practice, although it's possible there's still
a race condition.

It's not a big deal if there's still a race condition: the old behaviour was equivalent
to losing the race with 100% probability, and now it's much closer to 0%.